fix(migration): extracting prop removal into a helper.
